requirements-dev.txt
# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
ruff>=0.1.0

# Runtime dependencies (needed for import during tests)
pillow>=10.0.0
